+++ refinedstorage-rei-integration-neoforge/src/main/java/com/refinedmods/refinedstorage/rei/forge/TransferInputType.java
@@ -1,0 +1,7 @@
+package com.refinedmods.refinedstorage.rei.forge;
+
+enum TransferInputType {
+    AVAILABLE,
+    MISSING,
+    AUTOCRAFTABLE
+}