+++ CHANGELOG.md
@@ -7,6 +7,17 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.4.0] - 2024-10-12
+
+### Added
+
+-   The recipe transfer support in the Pattern Grid and Crafting Grid now indicates whether all or some resources are autocraftable.
+-   You can now start autocrafting for missing autocraftable items from the Crafting Grid via the recipe transfer.
+
+### Fixed
+
+-   Support for Refined Storage v2.0.0-milestone.4.8.
+
 ## [0.3.1] - 2024-08-11
 
 ### Fixed
@@ -69,7 +80,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   Support for REI collapsable entries.
 -   Grid / REI search box synchronization options.
 
-[Unreleased]: https://github.com/refinedmods/refinedstorage-rei-integration/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/refinedmods/refinedstorage-rei-integration/compare/v0.4.0...HEAD
+
+[0.4.0]: https://github.com/refinedmods/refinedstorage-rei-integration/compare/v0.3.1...v0.4.0
 
 [0.3.1]: https://github.com/refinedmods/refinedstorage-rei-integration/compare/v0.3.0...v0.3.1